refactor: Создала отдельный компонент LayersItem
