+++ src/entities/LayersItem/index.ts
@@ -1,0 +1,1 @@
+export { LayersItem } from "./ui";
