+++ src/features/Layers/model/slice.ts
@@ -1,7 +1,8 @@
 import { PayloadAction, createSlice } from "@reduxjs/toolkit";
 import { LayersState } from "./types";
-import { ILayer } from "../ui/types";
+
 import { setProjectLayersThunk } from "./layersThunk";
+import { ILayer } from "entities/LayersItem/ui/types";
 
 const initialState: LayersState = {
   layers: [],
