+++ src/features/Layers/model/types.ts
@@ -1,4 +1,4 @@
-import { ILayer } from "features/Layers/ui/types";
+import { ILayer } from "entities/LayersItem/ui/types";
 
 export type LayersState = {
   layers: ILayer[];
