+++ src/features/Layers/ui/index.tsx
@@ -1,32 +1,15 @@
-import {
-  createStyles,
-  UnstyledButton,
-  Text,
-  Group,
-  ActionIcon,
-  Tooltip,
-  getStylesRef,
-  Stack,
-  rem,
-  ScrollArea,
-} from "@mantine/core";
-import { IconTrash, IconLayersSubtract, IconEye, IconEyeOff, IconPlus } from "@tabler/icons-react";
+import { createStyles, Text, Group, ActionIcon, Tooltip, Stack, ScrollArea } from "@mantine/core";
+import { IconPlus } from "@tabler/icons-react";
 import { useAppDispatch, useAppSelector } from "app/store/hooks";
 import { SliderHover } from "shared/SliderHover/ui";
-import {
-  addLayer,
-  deleteLayer,
-  setActiveLayer,
-  toggleVisibility,
-  changeLayerLabel,
-  changeLayerOpacity,
-} from "../model/slice";
-import { ILayer } from "./types";
-import { EditableText } from "shared/EditableText";
+import { addLayer, setActiveLayer, changeLayerOpacity } from "../model/slice";
+
 import { useFirebaseDb } from "shared/hooks";
 import { useListState } from "@mantine/hooks";
 import { DragDropContext, Droppable, Draggable } from "react-beautiful-dnd";
 import { useEffect } from "react";
+import { LayersItem } from "entities/LayersItem";
+import { ILayer } from "entities/LayersItem/ui/types";
 
 const useStyles = createStyles((theme) => ({
   item: {
@@ -35,31 +18,6 @@ const useStyles = createStyles((theme) => ({
 
   itemDragging: {
     boxShadow: theme.shadows.sm,
-  },
-
-  symbol: {
-    fontSize: rem(30),
-    fontWeight: 700,
-    width: rem(60),
-  },
-  layer: {
-    "&:hover": {
-      backgroundColor: theme.colors.gray[0],
-      color: theme.black,
-    },
-  },
-
-  layerActive: {
-    "&, &:hover": {
-      backgroundColor: theme.fn.variant({
-        variant: "light",
-        color: theme.primaryColor,
-      }).background,
-      color: theme.fn.variant({ variant: "light", color: theme.primaryColor }).color,
-      [`& .${getStylesRef("icon")}`]: {
-        color: theme.fn.variant({ variant: "light", color: theme.primaryColor }).color,
-      },
-    },
   },
 }));
 
@@ -71,41 +29,7 @@ export function Layers() {
   const [state, handlers] = useListState(layers);
   const active = useAppSelector((state) => state.layers.activeLayer);
   const projectId = useAppSelector((state) => state.projects.openProjectId);
-  const { updateProjectLayer, deleteProjectLayer } = useFirebaseDb();
-
-  const handleEditableTextChange = (layer: ILayer, event: React.ChangeEvent<HTMLInputElement>) => {
-    const newLabel = event.currentTarget.value;
-    const layerProps = { ...layer };
-    dispatch(
-      changeLayerLabel({
-        id: event.currentTarget.id,
-        newLabel,
-      })
-    );
-    layerProps.label = newLabel;
-    updateProjectLayer({
-      projectId,
-      layer: layerProps,
-    });
-  };
-
-  const toggleVisibilityHandler = (layer: ILayer) => {
-    const layerProps = { ...layer };
-    layerProps.isVisible = !layer.isVisible;
-    dispatch(toggleVisibility(layerProps));
-    updateProjectLayer({
-      projectId,
-      layer: layerProps,
-    });
-  };
-
-  const deleteLayerHandler = (layer: ILayer) => {
-    dispatch(deleteLayer(layer.id));
-    deleteProjectLayer({
-      projectId,
-      layer,
-    });
-  };
+  const { updateProjectLayer } = useFirebaseDb();
 
   const onChangeOpacityLayerHandler = (value: number) => {
     if (active?.id) {
@@ -150,52 +74,7 @@ export function Layers() {
           {...provided.dragHandleProps}
           ref={provided.innerRef}
         >
-          <Group
-            key={layer.id}
-            position="apart"
-            fz="xs"
-            fw="500"
-            p="xs"
-            w="100%"
-            className={cx(classes.layer, {
-              [classes.layerActive]: layer.id === active?.id,
-            })}
-            onClick={() => {
-              dispatch(setActiveLayer(layer));
-            }}
-          >
-            <Group>
-              <IconLayersSubtract size={20} stroke={1.5} />
-              <EditableText
-                id={`${layer.id}`}
-                text={layer.label}
-                handleChange={(event) => handleEditableTextChange(layer, event)}
-              />
-            </Group>
-            <div>
-              <Tooltip label="Видимость слоя" withArrow position="bottom">
-                <UnstyledButton
-                  mr="xs"
-                  data-testid="toggle-visibility-button"
-                  onClick={(e) => {
-                    toggleVisibilityHandler(layer);
-                    e.stopPropagation();
-                  }}
-                >
-                  {layer.isVisible ? (
-                    <IconEye size={20} stroke={1.5} />
-                  ) : (
-                    <IconEyeOff size={20} stroke={1.5} />
-                  )}
-                </UnstyledButton>
-              </Tooltip>
-              <Tooltip label="Удалить слой" withArrow position="bottom">
-                <UnstyledButton onClick={() => deleteLayerHandler(layer)}>
-                  <IconTrash size={20} stroke={1.5} />
-                </UnstyledButton>
-              </Tooltip>
-            </div>
-          </Group>
+          <LayersItem layer={layer} />
         </div>
       )}
     </Draggable>
